Support for AoN, Stop, Market orders

+++ MiscClients/spring_reactjs/src/main/java/org/DistributedATS/WebTraderRest/quickfix/Position.java
@@ -32,23 +32,24 @@ import java.io.Serializable;
 
 public class Position implements Serializable {
 	
-	public Double buy_amt = 0.0;
-	public Double sell_amt = 0.0;
-	public Double buy_avg_price = 0.0;
-	public Double sell_avg_price = 0.0;
+	// All VWAP members in ticks
+	public Integer buy_amt = 0;
+	public Integer sell_amt = 0;
+	public Integer buy_avg_price = 0;
+	public Integer sell_avg_price = 0;
 	
-	public Double getPosition()
+	public Integer getPosition()
 	{
 		return (this.buy_amt - this.sell_amt);
 	}
 	
-	public Double getVWAP()
+	public Integer getVWAP()
 	{
 		if ( this.buy_amt-this.sell_amt != 0 )
 		{
 			return (this.buy_amt*this.buy_avg_price - this.sell_amt*this.sell_avg_price)/(this.buy_amt-this.sell_amt);
 		} else {
-			return 0.0;
+			return 0;
 		}
 	};
